refactor: move utils to separate file
